requirements.txt
pyautogui
pytest
selenium
pytest_bdd
allure-pytest
openpyxl
pyodbc
